DataFlash: change "unsigned long" to "uint32_t" in callback functions

+++ libraries/DataFlash/DataFlash_APM1.cpp
@@ -317,7 +317,7 @@ void DataFlash_APM1::BlockErase (uint16_t BlockAdr)
 
 
 
-void DataFlash_APM1::ChipErase(void (*delay_cb)(unsigned long))
+void DataFlash_APM1::ChipErase(void (*delay_cb)(uint32_t))
 {
 
   dataflash_CS_active();     // activate dataflash command decoder
